Move functions to where they ware used an mark them private
